Reorganized Components to have sperate panel and button folders.
