+++ src/components/panels/CenterPanel.js
