+++ src/components/panels/OutlinedPanel.js
