+++ src/components/panels/index.js
